+++ src/Transformers/SpatieEnumTransformer.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace Spatie\TypeScriptTransformer\Transformers;
+
+use ReflectionClass;
+use Spatie\Enum\Enum;
+use Spatie\TypeScriptTransformer\Structures\TransformedType;
+
+class SpatieEnumTransformer implements Transformer
+{
+    public function canTransform(ReflectionClass $class): bool
+    {
+        return $class->isSubclassOf(Enum::class);
+    }
+
+    public function transform(ReflectionClass $class, string $name): TransformedType
+    {
+        return TransformedType::create(
+            $class,
+            $name,
+            "export type {$name} = {$this->resolveOptions($class)};"
+        );
+    }
+
+    private function resolveOptions(ReflectionClass $class): string
+    {
+        /** @var \Spatie\Enum\Enum $enum */
+        $enum = $class->getName();
+
+        $options = array_map(
+            fn ($enum) => "'{$enum}'",
+            array_keys($enum::toArray())
+        );
+
+        return implode(' | ', $options);
+    }
+}