+++ CHANGELOG.md
@@ -5,3 +5,8 @@ All notable changes to `typescript-transformer` will be documented in this file
 ## 1.0.0 - 2020-09-02
 
 - initial release
+
+## 1.0.1 - 2020-11-26
+
+- Fix some capitalization in namespace names
+- Added `SpatieEnumTransformer` from the `laravel-typescript-transformer` package